% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Div_Stats_Map.R
\name{Div_Stats_Map}
\alias{Div_Stats_Map}
\title{\bold{WARNING! This function has been deprecated and is no longer supported. Please use the Point_map function instead.}
A function to map diversity statistics.}
\usage{
Div_Stats_Map(
  dat,
  plot.type = "all",
  statistic,
  breaks = NULL,
  col,
  Lat_buffer = 1,
  Long_buffer = 1,
  write = FALSE,
  prefix = NULL
)
}
\arguments{
\item{dat}{Data frame or character string that supplies the input data. If it is a character string, the file should be a csv. The first column should be the statistic to be plotted and named the same as the statistic argument. The second column is Population indicating which population each row belongs to. The third column is the standard deviation, the fourth column is Long indicating the longitude, and the fifth column is Lat, indicating the latitude.}

\item{plot.type}{Character string. Options are all, point, or interpolated. All is recommended and will generate a map with points colored according to heterozygosity as well as a rater of interpolated heterozygosity values.}

\item{statistic}{Character string. The statistic to be plotted.}

\item{breaks}{Numeric. The breaks used to generate the color ramp when plotting. Users should supply 3 values if custom breaks are desired.}

\item{col}{Character vector indicating the colors you wish to use for plotting, three colors are allowed (low, mid, high). The first color will be the low color, the second the middle, the third the high.}

\item{Lat_buffer}{Numeric. A buffer to customize visualization.}

\item{Long_buffer}{Numeric. A buffer to customize visualization.}

\item{write}{Boolean. Whether or not to write the output to a file in the current working directory.}

\item{prefix}{Character string that will be appended to file output.}
}
\value{
A list containing maps and the data frames used to generate them.
}
\description{
\bold{WARNING! This function has been deprecated and is no longer supported. Please use the Point_map function instead.}
A function to map diversity statistics.
}
\examples{
\donttest{
data(Het_dat)
Test_het <- Div_Stats_Map(dat = Het_dat, plot.type = 'all',
statistic = "Heterozygosity",
Lat_buffer = 1, Long_buffer = 1, write = FALSE, prefix = 'Test_het')}
}
