% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HornedLizard_VCF}
\alias{HornedLizard_VCF}
\title{A vcfR object to be used in \code{Heterozygosity} and \code{Differentiation}.}
\format{
A vcfR object
\describe{
\item{vcfR object}{A vcfR object containing genotype and sample informaiton for 72 individuals.}

...
}
}
\source{
Farleigh, K., Vladimirova, S. A., Blair, C., Bracken, J. T., Koochekian, N., Schield, D. R., ... & Jezkova, T. (2021). The effects of climate and demographic history in shaping genomic variation across populations of the Desert Horned Lizard (Phrynosoma platyrhinos). Molecular Ecology, 30(18), 4481-4496.
}
\usage{
data(HornedLizard_Pop)
}
\description{
Data frame containing 4 columns and 72 rows
}
\examples{
 \donttest{
data("HornedLizard_Pop")
data("HornedLizard_VCF")
Test <- Heterozygosity(data = HornedLizard_VCF, pops = HornedLizard_Pop, write = FALSE)}


}
\keyword{datasets}
