% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Q_dat}
\alias{Q_dat}
\title{A list representing a q-matrix and the locality information associated with the qmatrix}
\format{
A list with two elements:
\describe{
\item{Qmat}{A q-matrix with 6 columns and 30 rows, the first column lists the sample name and the remaining 5 represent the contribution a genetic cluster to that individuals ancestry}
\item{Loc_dat}{The locality information for each individual in the q-matrix}

...
}
}
\source{
Data was generated by package authors.
}
\usage{
data(Q_dat)
}
\description{
List with two elements
}
\examples{
\donttest{
data(Q_dat)
Qmat <- Q_dat[[1]]
rownames(Qmat) <- Qmat[,1]
Loc <- Q_dat[[2]]
Test_all <- Ancestry_barchart(anc.mat = Qmat, pops = Loc, K = 5,
plot.type = 'all',col = c('#d73027', '#fc8d59', '#e0f3f8', '#91bfdb', '#4575b4'))}

}
\keyword{datasets}
