% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Div_Stats.R
\name{Div_stats}
\alias{Div_stats}
\title{\bold{WARNING! This function has been deprecated and is no longer supported. Please use the Heterozygosity and Private.alleles functions.} A function to estimate heterozygosity and the number of private alleles from a vcf file.}
\usage{
Div_stats(VCF, pops, ploidy, write = FALSE, prefix)
}
\arguments{
\item{VCF}{Character string indicating the name of the vcf file to be used in analysis.}

\item{pops}{Character string indicating the name of the population assignment file. This file should have four columns and be in the same order as your vcf file. The first column named Sample indicates the sample name. The second column named Population indicates the population assignment of each individual. The third column named Longitude indicates the longitude of the sample.  The fourth column named Latitude indicates the latitude of the sample.}

\item{ploidy}{Numeric. The ploidy of the data.}

\item{write}{Boolean. Whether or not to write the output to a file in the current working directory.}

\item{prefix}{Character string that will be appended to file output.}
}
\value{
A list containing the estimated diversity statistics, model output from linear regression of these statistics against latitude, and model plots.
}
\description{
\bold{WARNING! This function has been deprecated and is no longer supported. Please use the Heterozygosity and Private.alleles functions.} A function to estimate heterozygosity and the number of private alleles from a vcf file.
}
\examples{
\donttest{
data("HornedLizard_Pop")
data("HornedLizard_VCF")
Test <- Div_stats(VCF = HornedLizard_VCF, pops = HornedLizard_Pop,
ploidy = 2, write = FALSE)}
}
