% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Het_dat}
\alias{Het_dat}
\title{A data frame of hypothetical heterozygosity data produced by Div_Stats.}
\format{
A data frame with 5 columns and 3 rows:
\describe{
\item{Heterozygosity}{Estimated heterozygosity}
\item{Pop}{Population assignment}
\item{Standard.Deviation}{standard deviation}
\item{Long}{Longitude}
\item{Lat}{Latitude}

...
}
}
\source{
Coordinates and population names taken from Farleigh, K., Vladimirova, S. A., Blair, C., Bracken, J. T., Koochekian, N., Schield, D. R., ... & Jezkova, T. (2021). The effects of climate and demographic history in shaping genomic variation across populations of the Desert Horned Lizard (Phrynosoma platyrhinos). Molecular Ecology, 30(18), 4481-4496.
}
\usage{
data(Het_dat)
}
\description{
Data frame containing 5 columns and 3 rows
}
\examples{
data(Het_dat)
Test_het <- Div_Stats_Map(dat = Het_dat, plot.type = 'point',
statistic = "Heterozygosity", countries = c('united states of america', 'mexico'),
Lat_buffer = 1, Long_buffer = 1, prefix = 'Test_het')

}
\keyword{datasets}
