% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heterozygosity.R
\name{Heterozygosity}
\alias{Heterozygosity}
\title{A function to estimate heterozygosity.}
\usage{
Heterozygosity(
  data,
  pops,
  statistic,
  write = FALSE,
  prefix,
  population_col = NULL,
  individual_col = NULL
)
}
\arguments{
\item{data}{Character. String indicating the name of the vcf file or vcfR object to be used in the analysis.}

\item{pops}{Character. String indicating the name of the population assignment file or dataframe containing the population assignment information for each individual in the data. This file must be in the same order as the vcf file and include columns specifying the individual and the population that individual belongs to. The first column should contain individual names and the second column should indicate the population assignment of each individual. Alternatively, you can indicate the column containing the individual and population information using the individual_col and population_col arguments.}

\item{statistic}{Character. String indicating the statistic to calculate. Options are ...}

\item{write}{Boolean. Whether or not to write the output to a file in the current working directory.}

\item{prefix}{Character. String that will be appended to file output.}

\item{population_col}{Numeric. Optional argument (a number) indicating the column that contains the population assignment information.}

\item{individual_col}{Numeric. Optional argument (a number) indicating the column that contains the individuals (i.e., sample name) in the data.}
}
\value{
A list containing the estimated diversity statistics, model output from linear regression of these statistics against latitude, and model plots.
}
\description{
A function to estimate heterozygosity.
}
\examples{
\donttest{
data("HornedLizard_Pop")
data("HornedLizard_VCF")
Test <- Div_stats(VCF = HornedLizard_VCF, pops = HornedLizard_Pop,
ploidy = 2, write = FALSE)}
}
