% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heterozygosity.R
\name{Heterozygosity}
\alias{Heterozygosity}
\title{A function to estimate heterozygosity.}
\usage{
Heterozygosity(
  data,
  pops,
  ploidy,
  statistic,
  write = FALSE,
  prefix,
  population_col = NULL,
  individual_col = NULL
)
}
\arguments{
\item{pops}{Character string indicating the name of the population assignment file. This file should have four columns and be in the same order as your vcf file. The first column named Sample indicates the sample name. The second column named Population indicates the population assignment of each individual. The third column named Longitude indicates the longitude of the sample.  The fourth column named Latitude indicates the latitude of the sample.}

\item{ploidy}{Numeric. The ploidy of the data.}

\item{write}{Boolean. Whether or not to write the output to a file in the current working directory.}

\item{prefix}{Character string that will be appended to file output.}

\item{VCF}{Character string indicating the name of the vcf file to be used in analysis.}
}
\value{
A list containing the estimated diversity statistics, model output from linear regression of these statistics against latitude, and model plots.
}
\description{
A function to estimate heterozygosity.
}
\examples{
\donttest{
data("HornedLizard_Pop")
data("HornedLizard_VCF")
Test <- Div_stats(VCF = HornedLizard_VCF, pops = HornedLizard_Pop,
ploidy = 2, write = FALSE)}
}
