% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Point_map.R
\name{Point_map}
\alias{Point_map}
\title{\bold{WARNING! This function has been deprecated and is no longer supported. Please use the Point_map function instead.}
A function to map diversity statistics.}
\usage{
Point_map(dat, statistic, breaks = NULL, col, Lat_buffer = 1, Long_buffer = 1)
}
\arguments{
\item{dat}{Data frame or character string that supplies the input data. If it is a character string, the file should be a csv. The first column should be the statistic to be plotted. The coordinates of each row should be indicated by columns named Longitude and Latitude.}

\item{statistic}{Character string. The statistic to be plotted.}

\item{breaks}{Numeric. The breaks used to generate the color ramp when plotting. Users should supply 3 values if custom breaks are desired.}

\item{col}{Character vector indicating the colors you wish to use for plotting, three colors are allowed (low, mid, high). The first color will be the low color, the second the middle, the third the high.}

\item{prefix}{Character string that will be appended to file output.}

\item{Long_col}{Numeric. A number indicating which column contains the longitude information.}

\item{Lat_col}{Numeric. A number indicating which column contains the latitude information.}

\item{write}{Boolean. Whether or not to write the output to a file in the current working directory.}
}
\value{
A list containing maps and the data frames used to generate them.
}
\description{
\bold{WARNING! This function has been deprecated and is no longer supported. Please use the Point_map function instead.}
A function to map diversity statistics.
}
\examples{
\donttest{
data(Het_dat)
Test <- Point_map(Het_dat, statistic = "Heterozygosity")}
}
